+++ classifai-core/pom.xml
@@ -20,7 +20,7 @@
     <parent>
         <groupId>ai.classifai</groupId>
         <artifactId>classifai</artifactId>
-        <version>1.0-SNAPSHOT</version>
+        <version>1.0.0</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
