+++ classifai-core/src/main/java/ai/classifai/database/annotation/AnnotationQuery.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package ai.classifai.database.annotationdb;
+package ai.classifai.database.annotation;
 
 /**
  * Common SQL Query for Each Annotation Type
@@ -26,9 +26,11 @@ public abstract class AnnotationQuery
 
     protected final static String RETRIEVE_DATA_PATH = "select img_path from Project where uuid = ? and project_id = ?";
 
-    protected final static String DELETE_DATA = "delete from Project where uuid = ? and project_id = ?";
+    protected final static String LOAD_VALID_PROJECT_UUID = "select img_path from Project where project_id = ? and uuid = ?";
 
-    protected final static String LOAD_VALID_PROJECT_UUID = "load valid project uuid";
+    protected final static String DELETE_PROJECT_UUID_LIST = "delete from Project where project_id = ?";
+
+    protected final static String DELETE_PROJECT_UUID = "delete from Project where project_id = ? and uuid = ?";
 
     public static String createData() { return CREATE_DATA; }
 
@@ -36,5 +38,7 @@ public abstract class AnnotationQuery
 
     public static String loadValidProjectUUID() { return LOAD_VALID_PROJECT_UUID; }
 
-    public static String deleteData() { return DELETE_DATA; }
+    public static String deleteProjectUUIDList() { return DELETE_PROJECT_UUID_LIST; }
+
+    public static String deleteProjectUUID() { return DELETE_PROJECT_UUID; }
 }
