+++ classifai-core/src/main/java/ai/classifai/database/annotation/boundingboxdb/BoundingBoxDbServiceable.java
