+++ classifai-core/src/main/java/ai/classifai/database/annotation/segdb/SegDbServiceable.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package ai.classifai.database.segdb;
+package ai.classifai.database.annotation.segdb;
 
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonObject;
