+++ classifai-core/src/main/java/ai/classifai/database/loader/ProjectLoader.java
@@ -191,4 +191,21 @@ public class ProjectLoader
         fileSystemStatus = status;
     }
 
+    /***
+     * Delete UUID from list
+     *
+     * @param uuid
+     * @return false if uuid in sanityUUIDLis not found
+     */
+    public boolean deleteUUID(Integer uuid)
+    {
+        if(sanityUUIDList.contains(uuid))
+        {
+            sanityUUIDList.remove(uuid);
+
+            return true;
+        }
+        return false;
+    }
+
 }
