+++ classifai-uberjar/pom.xml
@@ -19,7 +19,7 @@
     <parent>
         <groupId>ai.classifai</groupId>
         <artifactId>classifai</artifactId>
-        <version>1.0-SNAPSHOT</version>
+        <version>1.0.0</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
@@ -33,12 +33,12 @@
                 <dependency>
                     <groupId>ai.classifai</groupId>
                     <artifactId>classifai-core</artifactId>
-                    <version>1.0-SNAPSHOT</version>
+                    <version>1.0.0</version>
                 </dependency>
                 <dependency>
                     <groupId>ai.classifai</groupId>
                     <artifactId>classifai-api</artifactId>
-                    <version>1.0-SNAPSHOT</version>
+                    <version>1.0.0</version>
                 </dependency>
             </dependencies>
