+++ startClassifai.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-CLASSIFAI_VERSION="1.0-SNAPSHOT"
+CLASSIFAI_VERSION="1.0.0"
 CLASSIFAI_JAR="classifai-uberjar-$CLASSIFAI_VERSION-dev.jar"
 
 export CLASSIFAI_ABSPATH_JAR="$HOME/.m2/repository/ai/classifai/classifai-uberjar/$CLASSIFAI_VERSION/$CLASSIFAI_JAR"
